Version: 2013-05-03 ksh93v-
Sourced from:
53ca97d5b6e30fc73f728b3897310abf15b30b349ba492356bf455552fa7b735  ast-open.2013-04-22.tgz
16df539e552e8c2919d597dfcd9a4475482975d394fedfb5cc30d5478b67b59b  INIT.2013-04-22.tgz
87b1a0fe4b6ccd48b148e39a029a813f4922044684878ed162500d75da771642  ast-ksh.2013-05-03.tgz

+++ src/lib/libast/misc/spawnvex.c
@@ -520,7 +520,7 @@ spawnvex_add(Spawnvex_t* vex, intmax_t op, intmax_t arg, Spawnvex_f callback, vo
 			return -1;
 		}
 	}
-	switch (op)
+	switch ((int)op)
 	{
 	case SPAWN_frame:
 		arg = vex->frame;
